Update Time Converter.java

+++ edhesive/Time Converter.java
@@ -10,18 +10,18 @@ class Main{
         
         System.out.println("Seconds Since Unix Epoch:");
         
-        long sec = s.nextLong();
+        int sec = s.nextInt();
         
-        long yr = (sec / 31104000);
+        int yr = (sec / 31104000);
         sec -= yr * 31104000;
         yr += 1970;
-        long mon = sec / 2592000;
+        int mon = sec / 2592000;
         sec -= mon * 2592000;
-        long day = sec / 86400;
+        int day = sec / 86400;
         sec -= day * 86400;
-        long hr = sec / 3600;
+        int hr = sec / 3600;
         sec -= hr * 3600;
-        long min = sec / 60;
+        int min = sec / 60;
         sec -= min * 60;
         
         System.out.println("Year: " + yr);
@@ -36,3 +36,5 @@ class Main{
     }
 
 }
+
+
